Hurray, it works! Almost the final stage for its first version
The logic has been changed to random audio in .ogg format which plays it into a voice message, while the audio files themselves are pulled from the soundboards folder. I have hidden the contents of this folder, it is still a private bot. If someone suddenly finds it and wants to test it, just fill this folder with the content. I will describe this in the instructions later.